Adjust getParticiples and getClauses helpers
