+++ packages/yoastseo/src/languageProcessing/helpers/passiveVoice/directPrecedenceException.js
@@ -5,19 +5,19 @@ import getWords from "../word/getWords";
  * Checks whether the participle is directly preceded by a word from the direct precedence exception list.
  * If this is the case, the sentence part is not passive.
  *
- * @param {string} sentencePart The sentence part that contains the participle.
+ * @param {string} clause The sentence part that contains the participle.
  * @param {string} participle   The participle.
  * @param {Array} cannotDirectlyPrecedePassiveParticipleList List of words which cannot directly precede a passive participle.
  *
  * @returns {boolean} Returns true if a word from the direct precedence exception list is directly preceding
  *                    the participle, otherwise returns false.
  */
-export default function( sentencePart, participle, cannotDirectlyPrecedePassiveParticipleList = [] ) {
+export default function( clause, participle, cannotDirectlyPrecedePassiveParticipleList = [] ) {
 	// Break the sentence part up into words and convert to lower case.
-	const wordsInSentencePart = getWords( sentencePart ).map( word => word.toLowerCase() );
+	const wordsInClause = getWords( clause ).map( word => word.toLowerCase() );
 
 	// Search the participle in the word list.
-	const participleIndex = wordsInSentencePart.indexOf( participle.toLowerCase() );
+	const participleIndex = wordsInClause.indexOf( participle.toLowerCase() );
 
 	/*
 	 * There can be no exception in the following situations:
@@ -29,7 +29,7 @@ export default function( sentencePart, participle, cannotDirectlyPrecedePassiveP
 		return false;
 	}
 
-	const wordPrecedingParticiple = wordsInSentencePart[ participleIndex - 1 ];
+	const wordPrecedingParticiple = wordsInClause[ participleIndex - 1 ];
 
 	// Check if the word preceding the participle is in the exceptions list.
 	return includes( cannotDirectlyPrecedePassiveParticipleList, wordPrecedingParticiple );
